trying to set up netlify cms
