+++ package.json
@@ -21,13 +21,12 @@
     "@11ty/eleventy": "^0.11.1",
     "@11ty/eleventy-plugin-rss": "^1.1.0",
     "@11ty/eleventy-plugin-syntaxhighlight": "^3.0.6",
+    "@11ty/eleventy-img": "^0.3.0",
     "markdown-it": "^8.4.2",
     "markdown-it-anchor": "^5.2.5",
     "clean-css": "^4.2.3",
-    "dotenv": "^8.2.0",
-    "lighthouse-ci": "^1.11.0"
+    "dotenv": "^8.2.0"
   },
   "dependencies": {
-    "@11ty/eleventy-img": "^0.3.0"
   }
 }
