+++ src/_data/util.js
@@ -1,0 +1,28 @@
+let site = require("./site.json");
+const dotenv = require('dotenv').config()
+let now = new Date();
+const fs = require("fs"); // Or `import fs from "fs";` with ESM
+const path = require("path");
+module.exports = {
+    getFile: function(pathLoc){
+        //This gets the path of the file asked.  If you ask for header.njk
+        //it looks for ./src/_includes/layouts/header.njk //if found returns this
+        //if not found, looks for ./src/_includes/theme/*themename*/header.njk// if found returns this
+        //if neither found, console logs file not found
+        //assume they mean locally the _includes folder
+        let layoutsFolder = "./src/_includes/"
+        let themeFolder = "./src/_includes/theme/" + site.theme + "/"
+        let localFilePath = path.resolve(path.join(layoutsFolder, pathLoc))
+        let themeFilePath = path.resolve(path.join(themeFolder, pathLoc))
+        if(fs.existsSync(localFilePath)){
+            return localFilePath.toString();
+        } else if (fs.existsSync(themeFilePath)) {
+            return themeFilePath.toString();
+        } else {
+            console.log(`file not found ${pathLoc} in ${localFilePath} or ${themeFilePath}`)
+        }
+    },
+    environment: process.env.ELEVENTY_ENV,
+    lastBuildDate: now.toLocaleString('en-CA',{hour12:false, timeZone: 'America/Chicago'}).replace(',',''),
+    lastBuildYear: now.getFullYear(),
+}
