+++ src/_includes/theme/jace-docs/layouts/simple.njk
@@ -1,4 +1,4 @@
-{% extends site.getFile("layouts/base.njk") %}
+{% extends util.getFile("layouts/base.njk") %}
 {% block content %}
 {% if backgroundImage !="" %}
 <div class="page-title sp" style="background-image: url({{ backgroundImage }})">
