+++ src/_includes/theme/jace-simple/layouts/base.njk
@@ -58,8 +58,8 @@
 
 <!-- capture the CSS content as a Nunjucks variable -->
 {% set css %}
-      {% include site.getFile("flex.css") ignore missing %}
-      {% include site.getFile("prism-a11y-dark.css") ignore missing %}
+      {% include util.getFile("flex.css") ignore missing %}
+      {% include util.getFile("prism-a11y-dark.css") ignore missing %}
 {% endset %}
 <!-- feed it through our cssmin filter to minify -->
 <style>
@@ -72,7 +72,7 @@
 </head>
 
 <body> 
-  {% include site.getFile("layouts/header.njk") ignore missing %}
+  {% include util.getFile("layouts/header.njk") ignore missing %}
   
   <main>
     <article>
@@ -85,7 +85,7 @@
     </article>
     
   </main>
-  {% include site.getFile("layouts/footer.njk") ignore missing %}
+  {% include util.getFile("layouts/footer.njk") ignore missing %}
 </div>
 <script>
 var toggle = document.getElementById("theme-toggle");
