+++ src/_includes/theme/jace-simple/layouts/home.njk
@@ -1,4 +1,4 @@
-{% extends site.getFile("layouts/base.njk") %}
+{% extends util.getFile("layouts/base.njk") %}
 {%- block seo -%}
 <!--Twitter Card-->
 <meta name="twitter:card" content="summary">
