+++ src/_includes/theme/jace-ty/layouts/post.njk
@@ -1,19 +1,18 @@
 ---
 pagination:
   data: collections.post
-  size: 20
+  size: 2
   alias: posts
   reverse: true
-permalink: /post/{% if pagination.pageNumber > 0 %}/{{ pagination.pageNumber }}/{% endif %}index.html
+permalink: /posts/{% if pagination.pageNumber > 0 %}/{{ pagination.pageNumber }}/{% endif %}index.html
 ---
-{% extends site.getFile("layouts/simple.njk") %}
+{% extends util.getFile("layouts/simple.njk") %}
 
 {% block postContent %}
 
-
 <div class="posts-area">
 
-        {% for post in posts %}
+        {% for post in collections.post %}
         
         <article class="article">
           <img width="320px" class="article-thumb" data-sizes="auto" sizes="320px"
