+++ src/post.md
@@ -3,4 +3,7 @@ title: "Posts!"
 description: "All my posts"
 backgroundImage: "/assets/img/post-title.jpg"
 eleventyExcludeFromCollections: "true"
+permalink: /post/
+eleventyComputed:
+  layout: "/theme/{{ site.theme }}/layouts/post.njk"
 ---
