+++ src/post/post.11tydata.js
@@ -1,7 +1,9 @@
-let site = require("../_data/site.js");
+let site = require("../_data/site.json");
+console.log('in post.11tydata.js files site', site)
+console.log('site.theme', site.theme)
 module.exports = function() {
     var returnObj = {
-        "layout": `/theme/${site.theme}/layouts/post-single.njk`,
+        "layout": `theme/${site.theme}/layouts/post-single.njk`,
         "tags": "post",
         "imageName": "featured-640.webp",
         "imageWidth": "640",
